Move LockedByTag to a new folder
